pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "os-lem"
version = "0.0.1"
description = "Open scriptable lumped-element loudspeaker / enclosure simulator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
